configura endpoints e regra de negocio

+++ src/main/java/pt/com/hospital/api/Controller/PacienteController.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pt.com.hospital.application.DTO.mapper.PacienteMapper;
 import pt.com.hospital.application.DTO.request.PacienteRequest;
 import pt.com.hospital.application.DTO.response.PacienteResponse;
 import pt.com.hospital.domain.Service.PacienteService;
@@ -18,37 +17,37 @@ public class PacienteController {
     private PacienteService service;
 
     @PostMapping("/paciente/")
-    public ResponseEntity<PacienteResponse> save(@RequestBody PacienteRequest request){
+    public ResponseEntity<PacienteResponse> save(@RequestBody PacienteRequest request) {
         var create = service.createPaciente(request);
         return ResponseEntity.ok(create);
     }
 
     @GetMapping("/paciente/")
-    public  ResponseEntity<List<PacienteResponse>> findAll(){
-       var findPaciente = service.findAll();
-        return  ResponseEntity.ok(findPaciente);
+    public ResponseEntity<List<PacienteResponse>> findAll() {
+        var findPaciente = service.findAll();
+        return ResponseEntity.ok(findPaciente);
     }
 
     @GetMapping("/paciente/{id}")
-    public ResponseEntity<PacienteResponse> getPacienteById(@PathVariable long id, PacienteRequest request){
-        var create = PacienteMapper.toPaciente(request);
-        var getNovo = service.getPacienteById(id);
-        PacienteResponse getPaciente = PacienteMapper.toPacienteResponse(getNovo);
-        return  ResponseEntity.ok(getPaciente);
+    public ResponseEntity<PacienteResponse> getPacienteById(@PathVariable long id) {
+        var get = service.getPacienteById(id);
+        return ResponseEntity.ok(get);
     }
 
     @PutMapping("/paciente/{id}")
-    public ResponseEntity<PacienteResponse> update(@PathVariable long id,@RequestBody PacienteRequest request){
-        var update = PacienteMapper.toPaciente(request);
-        var newUpadte = service.updateById(id, update);
-        PacienteResponse updatePaciente = PacienteMapper.toPacienteResponse(newUpadte);
-        return new ResponseEntity<>(updatePaciente, HttpStatus.OK);
+    public ResponseEntity<PacienteResponse> update(@PathVariable long id, @RequestBody PacienteRequest request) {
+        var updatedById = service.updateById(id, request);
+        return new ResponseEntity<>(updatedById, HttpStatus.OK);
     }
 
     @DeleteMapping("/paciente/{id}")
     public ResponseEntity<Void> deleteById(@PathVariable long id){
-        service.deleteById(id);
-        return ResponseEntity.ok().build();
-    }
 
+        Boolean sucesso  = service.deleteById(id);
+        if (sucesso){
+            service.deleteById(id);
+            return  ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    }
 }
